chore(cli): extract watch loop, signature, and severity helpers
