+++ packages/cli/src/commands/diff.ts
@@ -18,6 +18,7 @@ import { addTransportOptions } from "../transport.js";
 import {
   CliExitError,
   handleErrors,
+  parseSeverity,
   readSnapshotFile,
   resolveFormat,
   stripAnsi,
@@ -26,22 +27,6 @@ import {
 import { sendWebhook } from "../webhook.js";
 import { captureSnapshot } from "./capture.js";
 import { executeCompositionDiff } from "./diff-composition.js";
-
-const VALID_SEVERITIES = new Set<string>(["safe", "warning", "breaking"]);
-
-/**
- * Validates that a string is a valid Severity level.
- *
- * @param value - The string to validate.
- * @param label - Label for the option (used in error messages).
- * @returns The validated Severity value.
- */
-function parseSeverity(value: string, label: string): Severity {
-  if (!VALID_SEVERITIES.has(value)) {
-    throw new Error(`Invalid ${label} value "${value}". Must be one of: safe, warning, breaking`);
-  }
-  return value as Severity;
-}
 
 /**
  * Resolves the "after" snapshot, either from file or by capturing from a live server.
